moving circle catch trial and rig_params added
added rig_params json which should allow individual rig flexibliity. after initial commit of it, this will be hidden so it doesn't overwrite. this controls for things like different com ports and in the future different solenoid water values.

also added catch trials to the moving circle task for better representation of catch trials... some plotting code may need to be changed to reflect this